image rotation
Images can be rotated to any degree clockwise or otherwise. We just need to define rotation matrix listing rotation point, degree of rotation and the scaling factor. 

The cv2.getRotationMatrix2D() function is used to create a rotation matrix for an image. It takes the following arguments:
The center of rotation for the image.
The angle of rotation in degrees.
The scale factor.
The cv2.warpAffine() function is used to apply a transformation matrix to an image. It takes the following arguments:
The python image to be transformed.
The transformation matrix.
The output image size.
The rotation angle can be positive or negative. A positive angle rotates the image clockwise, while a negative angle rotates the image counterclockwise.
The scale factor can be used to scale the image up or down. A scale factor of 1 will keep the image the same size, while a scale factor of 2 will double the size of the python image