image blurring
Image blurring is the technique of reducing the detail of an image by averaging the pixel values in the neighborhood. This can be done to reduce noise, soften edges, or make it harder to identify a picture. In many image processing tasks, image blurring is a common preprocessing step. It is useful in the optimization of algorithms such as image classification, object identification, and image segmentation. In OpenCV, a variety of different blurring methods are available, each with a particular trade-off between blurring strength and speed.

Some of the most common blurring techniques include:

Gaussian blurring: This is a popular blurring technique that uses a Gaussian kernel to smooth out the image.
Median blurring: This blurring technique uses the median of the pixel values in a neighborhood to smooth out the image.
Bilateral blurring: This blurring technique preserves edges while blurring the image.